# server/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpacetimeWiki REST API",
    "description": "Programmatic CRUD and search for pages, collections, and more. SpacetimeWiki is a real-time wiki with rich text editing, hierarchical collections, tagging, search, SSO/OAuth/LDAP auth, SCIM provisioning, WebAuthn/passkeys, import/export, and more.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/oauth/callback": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Oauth Callback",
        "description": "Handle OAuth callback.\n\nThe frontend sends the authorization code, code_verifier, and provider_id.\nThe backend exchanges the code for tokens and fetches user info.",
        "operationId": "oauth_callback_api_v1_auth_oauth_callback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthCallbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/login": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Oauth Login",
        "description": "Initiate OAuth login. Returns provider config for the frontend to build the redirect URL.\n\nThe frontend should:\n1. Call this endpoint with the provider_id\n2. Generate PKCE code_verifier + code_challenge\n3. Build the authorize URL and redirect the user",
        "operationId": "oauth_login_api_v1_auth_oauth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthLoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/oauth/user-links/{user_id}": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "List User Links",
        "description": "List all OAuth provider links for a user.",
        "operationId": "list_user_links_api_v1_auth_oauth_user_links__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OAuthUserLinkResponse"
                  },
                  "title": "Response List User Links Api V1 Auth Oauth User Links  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OAuthCallbackResponse": {
        "properties": {
          "user": {
            "additionalProperties": true,
            "type": "object",
            "title": "User",
            "description": "Authenticated wiki user object"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "OAuthCallbackResponse"
      },
      "OAuthLoginResponse": {
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/OAuthProviderResponse",
            "description": "Provider configuration for building redirect URL"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "OAuthLoginResponse"
      },
      "OAuthProviderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Provider ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name",
            "default": ""
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "URL slug",
            "default": ""
          },
          "provider_type": {
            "type": "string",
            "title": "Provider Type",
            "description": "Provider type: github, discord, slack, gitlab, generic",
            "default": ""
          },
          "authorize_url": {
            "type": "string",
            "title": "Authorize Url",
            "description": "OAuth authorize URL",
            "default": ""
          },
          "token_url": {
            "type": "string",
            "title": "Token Url",
            "description": "Token exchange URL",
            "default": ""
          },
          "userinfo_url": {
            "type": "string",
            "title": "Userinfo Url",
            "description": "User info URL",
            "default": ""
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "OAuth scope",
            "default": ""
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "OAuth client ID",
            "default": ""
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "Provider icon URL",
            "default": ""
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether provider is active",
            "default": true
          },
          "auto_register": {
            "type": "boolean",
            "title": "Auto Register",
            "description": "Auto-register new users",
            "default": false
          },
          "default_role": {
            "type": "string",
            "title": "Default Role",
            "description": "Default role for new users",
            "default": "member"
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "description": "Creator user ID",
            "default": ""
          },
          "created_at": {
            "type": "integer",
            "title": "Created At",
            "description": "Creation timestamp",
            "default": 0
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At",
            "description": "Last update timestamp",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "OAuthProviderResponse"
      },
      "OAuthUserLinkResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Link ID"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Wiki user ID"
          },
          "provider_id": {
            "type": "string",
            "title": "Provider Id",
            "description": "OAuth provider ID"
          },
          "external_id": {
            "type": "string",
            "title": "External Id",
            "description": "External user ID from provider",
            "default": ""
          },
          "external_username": {
            "type": "string",
            "title": "External Username",
            "description": "External username",
            "default": ""
          },
          "external_email": {
            "type": "string",
            "title": "External Email",
            "description": "External email",
            "default": ""
          },
          "token_expires_at": {
            "type": "integer",
            "title": "Token Expires At",
            "description": "Token expiry timestamp",
            "default": 0
          },
          "last_synced_at": {
            "type": "integer",
            "title": "Last Synced At",
            "description": "Last sync timestamp",
            "default": 0
          },
          "created_at": {
            "type": "integer",
            "title": "Created At",
            "description": "Creation timestamp",
            "default": 0
          },
          "updated_at": {
            "type": "integer",
            "title": "Updated At",
            "description": "Last update timestamp",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "provider_id"
        ],
        "title": "OAuthUserLinkResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "SpacetimeWiki API server"
    }
  ]
}